requirement.txt
alembic==1.0.10
Click==7.0
Flask==1.0.3
Flask-Migrate==2.5.0
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.10
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.3
Werkzeug==0.15.4
